refactor: renamed plugin files and directories in PascalCase for consistency
