+++ src/web/pmPlugins/MentionPlugin/removeMentionMarksIfSpansResized.ts
