+++ src/web/pmPlugins/MentionPlugin/stripPartialMentionMarks.ts
