+++ src/web/pmPlugins/MentionPlugin/subscribeMentionEvents.ts
