+++ src/web/pmPlugins/MergeAdjacentSameKindBlocksPlugin.ts
