+++ src/web/pmPlugins/ShortcutPlugin.ts
