+++ src/web/pmPlugins/StripBoldInStyledHeadingsPlugin.ts
