+++ src/web/pmPlugins/StripMarksOnImagePlugin.ts
